requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic>=2.6.4
pydantic-settings==2.0.3
httpx==0.24.1
python-dotenv==1.0.0
composio-core
python-multipart==0.0.6
